+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ v1.2.1
 ======
 Added
 -----
+* a new ``background`` formatoption has been implemented that allows to set the
+  facecolor of the axes (i.e. the background color for the plot)
 * compatibility for cartopy 0.18 (see `#14 <https://github.com/psyplot/psy-maps/pull/14>`__)
 * all plotmethods now have a ``transpose`` formatoption that can be used if the
   order of dimensions in the data is ``(x, y)`` rather than ``(y, x)``